Casting Fix
Fixed an issue where some functions called by gPhoton passed floats where ints were needed

+++ gPhoton/dbasetools.py
@@ -149,18 +149,18 @@ def get_valid_times(band, skypos, trange=None, detsize=1.1, verbose=0,
         for r in np.unique(
                 np.hstack([np.linspace(skypos[0],
                                        skypos[0]+skyrange[0]/2.,
-                                       np.ceil(skyrange[0]/(detsize/3.)),
+                                       int(np.ceil(skyrange[0]/(detsize/3.))),
                                        endpoint=True),
                            np.linspace(skypos[0], skypos[0]-skyrange[0]/2.,
-                                       np.ceil(skyrange[0]/(detsize/3.)),
+                                       int(np.ceil(skyrange[0]/(detsize/3.))),
                                        endpoint=True)])):
             for d in np.unique(
                     np.hstack([np.linspace(skypos[1],
                                            skypos[1]+skyrange[1]/2.,
-                                           np.ceil(skyrange[1]/(detsize/3.)),
+                                           int(np.ceil(skyrange[1]/(detsize/3.))),
                                            endpoint=True),
                                np.linspace(skypos[1], skypos[1]-skyrange[1]/2.,
-                                           np.ceil(skyrange[1]/(detsize/3.)),
+                                           int(np.ceil(skyrange[1]/(detsize/3.))),
                                            endpoint=True)])):
                 skypos_list += [[r, d]]
